feat(market-server-rust): a begining of data sanitazition on trade routes

+++ servers/rust-server/entity/wallet.rs
@@ -8,7 +8,8 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub bot_id: i32,
-    pub cash: i32,
+    #[sea_orm(column_type = "Decimal(Some((5, 2)))")]
+    pub cash: Decimal,
     #[sea_orm(column_type = "Decimal(Some((5, 2)))")]
     pub asset: Decimal,
     pub created_at: DateTimeWithTimeZone,
